+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   querying the underlying data.
 * Added new framework crates:
   - `CoreAudioTypes` / `objc2-core-audio-types`.
+  - `CoreGraphics` / `objc2-core-graphics`.
   - `CoreFoundation` / `objc2-core-foundation`.
   - `CoreVideo` / `objc2-core-video`.
   - `ScreenSaver` / `objc2-screen-saver`.